requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
bidict==0.22.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
comm==0.1.4
contourpy==1.1.1
csbdeep==0.7.4
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.0
fastapi==0.104.0
fastapi-socketio==0.0.10
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
ifaddr==0.2.0
imageio==2.31.6
ipykernel==6.26.0
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.5.0
jupyter_core==5.4.0
keras==2.14.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
llvmlite==0.41.1
Markdown==3.5
markdown2==2.4.10
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
multidict==6.0.4
nest-asyncio==1.5.8
networkx==3.2
nicegui==1.4.1
numba==0.58.1
numpy==1.26.1
nvidia-cublas-cu12==12.3.2.9
nvidia-cuda-nvrtc-cu12==12.3.52
nvidia-cuda-runtime-cu12==12.3.52
nvidia-cudnn-cu12==8.9.4.25
nvidia-tensorrt==99.0.0
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.9.10
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.24.4
pscript==0.7.7
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-multipart==0.0.6
python-socketio==5.10.0
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scipy==1.11.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
stardist==0.8.5
starlette==0.27.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
termcolor==2.3.0
tifffile==2023.9.26
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.22.0
uvloop==0.19.0
vbuild==0.8.2
watchfiles==0.21.0
wcwidth==0.2.8
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.2
